requirements.txt
appdirs==1.4.4
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.1
Django==3.1.5
django-filter==2.4.0
djangorestframework==3.12.2
filelock==3.0.12
httpie==2.3.0
idna==2.10
itypes==1.2.0
Jinja2==2.11.3
Markdown==3.3.3
MarkupSafe==1.1.1
psycopg2==2.8.6
Pygments==2.7.4
PySocks==1.7.1
pytz==2020.5
PyYAML==5.4.1
requests==2.25.1
requests-toolbelt==0.9.1
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.3
virtualenv==20.4.0
virtualenvwrapper-win==1.2.6
